pyproject.toml
[tool.poetry]
name = "tgofduty"
version = "0.1.0"
description = ""
authors = ["Dmitry Moskowski <me@corpix.ru>"]

[tool.poetry.dependencies]
python = "^3.9"
python-telegram-bot = "^13.9"
SQLAlchemy = "^1.4.29"
python-telegram-bot-calendar = "^1.0.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "pypi-public"
url = "https://pypi.org/simple/"
